move bucket lock/hold tests to correct directory
